+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Download at TODO
 * Moved de.larssh.utils.AetherUtils to de.larssh.utils.maven.AetherUtils
 * NumericTextComparator handles numerics as signed only if the sign is preceeded by a whitespace or at a strings start
 * ParseException initializes cause
+* Checkstyle warnings may be suppressed using @SuppressWarnings
 * Checkstyle suppressions file name changed from "suppressions.xml" to "checkstyle-suppressions.xml"
 * Checkstyle for unit tests suppresses checks ExecutableStatementCountCheck, MagicNumberCheck and MethodLengthCheck
 * Validate Javadoc inside Travis CI